+++ runners/flink/src/main/java/org/apache/beam/runners/flink/FlinkExecutionEnvironments.java
@@ -375,6 +375,13 @@ public class FlinkExecutionEnvironments {
                     : ExternalizedCheckpointCleanup.DELETE_ON_CANCELLATION);
       }
 
+      if (options.getUnalignedCheckpointEnabled()) {
+        flinkStreamEnv.getCheckpointConfig().enableUnalignedCheckpoints();
+      }
+      flinkStreamEnv
+          .getCheckpointConfig()
+          .setForceUnalignedCheckpoints(options.getForceUnalignedCheckpointEnabled());
+
       long minPauseBetweenCheckpoints = options.getMinPauseBetweenCheckpoints();
       if (minPauseBetweenCheckpoints != -1) {
         flinkStreamEnv